Verification stuff and updated readme.md

+++ SmartCity/api/views/bill.py
@@ -1,6 +1,7 @@
 from rest_framework.response import Response
 from rest_framework.decorators import api_view
 from stva.models import Bill
+from api.jwt import encode_token
 from ..serializers import BillSerializer
 
 
@@ -13,6 +14,11 @@ def get_all_bills(request):
 
 @api_view(["POST"])
 def add_bill(request):
+    if "accesToken" not in request.META.keys():
+        return Response(status=400) 
+    payload = jwt.encode_token(request.META["accessToken"])
+    if not jwt.verify(payload["expireDate"]):
+        return Response(status=401) 
     serializer = BillSerializer(data=request.data)
     if serializer.is_valid():
         serializer.save()
@@ -30,11 +36,13 @@ def get_bill_by_id(request, id):
 
 @api_view(["GET"])
 def get_bill_by_user(request):
-    
-    # TODO: Verify user by token from request first
-    
+    if "accesToken" not in request.META.keys():
+        return Response(status=400) 
+    payload = jwt.encode_token(request.META["accessToken"])
+    if not jwt.verify(payload["expireDate"]):
+        return Response(status=401) 
     try:
-        bills = Bill.objects.filter(owner=request.owner)    
+        bills = Bill.objects.filter(receiver=payload["email"])    
     except Bill.DoesNotExist:
         return Response(status=404)        
     serializer = BillSerializer(bills, many=True)
@@ -43,11 +51,15 @@ def get_bill_by_user(request):
 
 @api_view(["DELETE"])
 def delete_bill_by_id(request, id):
-    
-    # TODO: Verify user by token from request first
-    
+    if "accesToken" not in request.META.keys():
+        return Response(status=400) 
+    payload = jwt.encode_token(request.META["accessToken"])
+    if not jwt.verify(payload["expireDate"]):
+       return Response(status=401) 
     try:
         bill = Bill.objects.get(pk=id)    
+        if bill.receiver != payload["email"]:
+           return Response(status=401) 
     except Bill.DoesNotExist:
         return Response(status=404)   
     bill.delete()
